# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusMate AI Backend",
    "description": "大学生校园事务智能陪伴助手 — 后端 API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/knowledge/documents": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_knowledge_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummary"
                  },
                  "type": "array",
                  "title": "Response List Documents Api V1 Knowledge Documents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload Document",
        "description": "上传文档到知识库。\n\n使用 multipart/form-data 上传文件，元数据通过 Form 字段提供。\n上传的文档默认 is_demo=False(用户导入资料)。",
        "operationId": "upload_document_api_v1_knowledge_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_knowledge_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/documents/{document_id}": {
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_api_v1_knowledge_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/manage/{action}": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Data Management",
        "description": "数据管理操作。\n\n支持的 action:\n- delete_user_documents: 删除所有用户导入的知识库文档(保留测试环境资料)\n- delete_all_documents: 删除所有知识库文档(包括测试环境资料)\n\n备注:\n- 不再提供\"恢复演示资料 / 一键重置演示数据\"等生产接口;\n  测试环境资料仅在 dev/test 启动时通过 AUTO_IMPORT_DEMO 导入,\n  production 已被 config 校验拦截。",
        "operationId": "data_management_api_v1_knowledge_manage__action__post",
        "parameters": [
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataManagementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_knowledge_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "source_department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Department"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "applicable_students": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable Students"
          },
          "is_official": {
            "type": "boolean",
            "title": "Is Official",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_knowledge_documents_post"
      },
...
      "DataManagementResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "执行的清理动作"
          },
          "affected_count": {
            "type": "integer",
            "title": "Affected Count",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "action",
          "message"
        ],
        "title": "DataManagementResponse",
        "description": "数据清理操作响应。"
      },
      "DeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "document_id"
        ],
        "title": "DeleteResponse"
      },
      "DocumentSummary": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Department"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "applicable_students": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable Students"
          },
          "is_official": {
            "type": "boolean",
            "title": "Is Official"
          },
          "is_expired": {
            "type": "boolean",
            "title": "Is Expired"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo",
            "default": false
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          },
          "file_ext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Ext"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "content_hash",
          "is_official",
          "is_expired",
          "imported_at"
        ],
        "title": "DocumentSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}